Module wasm_run::prelude

The wasm-run Prelude

The purpose of this module is to alleviate imports of many common types:

use wasm_run::prelude::*;

Re-exports

pub use anyhow;

pub use async_std;

pub use cargo_metadata;

pub use fs_extra;

pub use futures;

pub use notify;

pub use tide;

pub use super::BuildArgs;

pub use super::BuildProfile;

pub use super::CargoChild;

pub use super::DefaultBuildArgs;

pub use super::DefaultServeArgs;

pub use super::Hooks;

pub use super::PackageExt;

pub use super::ServeArgs;

Structs

Metadata
Starting point for metadata returned by cargo metadata

Package
One or more crates described by a single Cargo.toml

Server
An HTTP server.

Enums

Message
A cargo message

Type Definitions

RecommendedWatcher
The recommended Watcher implementation for the current platform

Attribute Macros

main
Makes an entrypoint to your binary (not WASM).