Trait wasm_run::BuildArgs

pub trait BuildArgs: Downcast {
    fn build_path(&self) -> &PathBuf;

    fn profiling(&self) -> bool;

    fn default_build_path(&self) -> &PathBuf { ... }

    fn target_path(&self) -> &PathBuf { ... }

    fn metadata(&self) -> &Metadata { ... }

    fn package(&self) -> &Package { ... }

    fn run(self) -> Result<PathBuf>
    where
        Self: Sized + 'static,
    { ... }
}

A trait that allows overriding the build command.

Required methods

fn build_path(&self) -> &PathBuf

Build directory output.

fn profiling(&self) -> bool

Create a profiling build. Enable optimizations and debug info.

Loading content...
Provided methods

fn default_build_path(&self) -> &PathBuf

Default path for the build/public directory.

fn target_path(&self) -> &PathBuf

Path to the target directory.

fn metadata(&self) -> &Metadata

Metadata of the project.

fn package(&self) -> &Package

Package metadata.

fn run(self) -> Result<PathBuf> where
Self: Sized + 'static,

Run the build command.

Loading content...
Implementations

impl dyn BuildArgs

pub fn is<__T: BuildArgs>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: BuildArgs>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T: BuildArgs>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T: BuildArgs>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T: BuildArgs>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.