Trait wasm_run::PackageExt

pub trait PackageExt {
    fn cargo(&self, builder: impl FnOnce(&mut Command)) -> Result<CargoChild>;
}

An extension for Package and for Metadata to run a cargo command a bit more easily. Ideal for scripting.

Required methods

fn cargo(&self, builder: impl FnOnce(&mut Command)) -> Result<CargoChild>

Run the cargo command in the package's directory if ran on a Package or in the workspace root if ran on a Metadata.

Loading content...