Trait wasm_run::ServeArgs

pub trait ServeArgs: Downcast + Send {
    fn log(&self) -> bool;

    fn ip(&self) -> &str;

    fn port(&self) -> u16;

    fn build_args(&self) -> &dyn BuildArgs;

    fn run(self) -> Result<()>
    where
        Self: Sized + 'static,
    { ... }
}

A trait that allows overriding the serve command.

Required methods

fn log(&self) -> bool

Activate HTTP logs.

fn ip(&self) -> &str

IP address to bind.

Use 0.0.0.0 to expose the server to your network.

fn port(&self) -> u16

Port number.

fn build_args(&self) -> &dyn BuildArgs

Build arguments.

Loading content...
Provided methods

fn run(self) -> Result<()> where
Self: Sized + 'static,

Run the serve command.

Loading content...
Implementations

impl dyn ServeArgs

pub fn is<__T: ServeArgs>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: ServeArgs>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T: ServeArgs>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T: ServeArgs>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T: ServeArgs>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.